WIP function calls

+++ src/runtime/intfmt.c
@@ -32,7 +32,7 @@ int l_intfmt(lua_State* L) {
     }
     int is_unsigned = (nargs > 2) ? lua_toboolean(L, 3) : 0;
 
-    u64 uvalue = is_unsigned ? (u64)value : (u64)((value < 0) ? -value : value);
+    u64 uvalue = is_unsigned ? (u64)value : (u64)((value < 0) ? -(u64)value : value);
     do {
         // 0xdf (0b_1101_1111) normalizes the case of letters, i.e. 'A' => 'a'
         *--str = "0123456789abcdefghijklmnopqrstuvwxyz"[(uvalue % base) & 0xdf];
